+++ lib/mix/test/mix/tasks/deps.path_test.exs
@@ -39,7 +39,7 @@ defmodule Mix.Tasks.DepsPathTest do
     in_fixture "deps_status", fn ->
       Mix.Deps.Lock.write [raw_repo: "abcdef"]
       Mix.Tasks.Deps.Compile.run ["raw_repo"]
-      Mix.Tasks.Run.run ["Mix.shell.info", "RawRepo.hello"]
+      Mix.Tasks.Run.run ["-e", "Mix.shell.info RawRepo.hello"]
       assert_received { :mix_shell, :info, ["world"] }
     end
   after